requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pypdf
markdown
beautifulsoup4
pytest
httpx

